backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
email-validator==2.1.0

# Database & Storage (simplified for now)
sqlalchemy==2.0.23

# AI & API Integration
anthropic>=0.40.0
openai>=1.0.0

# WebSocket & Real-time
python-socketio[asyncio]==5.10.0
websockets==12.0

# Utilities & Helpers
python-dotenv==1.0.0
pyyaml>=6.0
requests>=2.31.0
aiofiles>=23.0.0
psutil>=5.9.0

# Development & Debugging
rich>=13.0.0